third_party/blink/web_tests/external/wpt/webrtc/RTCRtpReceiver-getSynchronizationSources.https-expected.txt

This is a testharness.js-based test.
PASS [audio] getSynchronizationSources() eventually returns a non-empty list
PASS [audio] RTCRtpSynchronizationSource.timestamp is a number
PASS [audio] RTCRtpSynchronizationSource.rtpTimestamp is a number [0, 2^32-1]
PASS [audio] getSynchronizationSources() does not contain SSRCs older than 10 seconds
PASS [audio] RTCRtpSynchronizationSource.timestamp is comparable to performance.timeOrigin + performance.now()
PASS [audio] RTCRtpSynchronizationSource.source is a number
PASS [video] getSynchronizationSources() eventually returns a non-empty list
PASS [video] RTCRtpSynchronizationSource.timestamp is a number
PASS [video] RTCRtpSynchronizationSource.rtpTimestamp is a number [0, 2^32-1]
PASS [video] getSynchronizationSources() does not contain SSRCs older than 10 seconds
PASS [video] RTCRtpSynchronizationSource.timestamp is comparable to performance.timeOrigin + performance.now()
PASS [video] RTCRtpSynchronizationSource.source is a number
PASS [audio-only] RTCRtpSynchronizationSource.audioLevel is a number [0, 1]
FAIL [audio-only] RTCRtpSynchronizationSource.voiceActivityFlag is a boolean assert_equals: expected "boolean" but got "undefined"
Harness: the test ran to completion.
